refactor: move `humanReadableVaule` to `src/logic/tokens/utils` directory
